Apply suggested fix to internal/ingest/sweeper_test.go from Copilot Autofix

+++ internal/ingest/sweeper_test.go
@@ -184,7 +184,7 @@ func TestFindGapSequence(t *testing.T) {
 		if seq <= 60 {
 			return &jetstream.RawStreamMsg{Time: cutoff.Add(-time.Duration(61-seq) * time.Minute)}, nil
 		}
-		return &jetstream.RawStreamMsg{Time: cutoff.Add(time.Duration(int64(seq-60)) * time.Minute)}, nil //nolint:gosec // test-only, values are small
+		return &jetstream.RawStreamMsg{Time: cutoff.Add(time.Duration(seq-60) * time.Minute)}, nil //nolint:gosec // test-only, values are small
 	}
 	// seqs 40..60 missing (already purged), <40 are old, >60 are within window.
 	sparseSequences := func(_ context.Context, seq uint64, _ ...jetstream.GetMsgOpt) (*jetstream.RawStreamMsg, error) {
